Run hooks in powershell on Windows
